feat(sdk): refactor into single messenger class
