+++ packages/sdk/src/index.ts
@@ -1,5 +1,4 @@
 export * from './interfaces'
 export * from './utils'
-export * from './cross-chain-provider'
 export * from './cross-chain-messenger'
 export * from './adapters'
