+++ packages/sdk/src/interfaces/index.ts
@@ -1,5 +1,4 @@
 export * from './bridge-adapter'
 export * from './cross-chain-messenger'
-export * from './cross-chain-provider'
 export * from './l2-provider'
 export * from './types'
