+++ packages/sdk/src/utils/coercion.ts
@@ -5,27 +5,32 @@ import {
   TransactionReceipt,
   TransactionResponse,
 } from '@ethersproject/abstract-provider'
+import { Signer } from '@ethersproject/abstract-signer'
 import { ethers, BigNumber } from 'ethers'
 
 import {
-  ProviderLike,
+  SignerOrProviderLike,
   TransactionLike,
   NumberLike,
   AddressLike,
 } from '../interfaces'
 
 /**
- * Converts a ProviderLike into a provider. Assumes that if the ProviderLike is a string then
- * it is a JSON-RPC url.
+ * Converts a SignerOrProviderLike into a Signer or a Provider. Assumes that if the input is a
+ * string then it is a JSON-RPC url.
  *
- * @param provider ProviderLike to turn into a provider.
- * @returns ProviderLike as a provider.
+ * @param signerOrProvider SignerOrProviderLike to turn into a Signer or Provider.
+ * @returns Input as a Signer or Provider.
  */
-export const toProvider = (provider: ProviderLike): Provider => {
-  if (typeof provider === 'string') {
-    return new ethers.providers.JsonRpcProvider(provider)
-  } else if (Provider.isProvider(provider)) {
-    return provider
+export const toSignerOrProvider = (
+  signerOrProvider: SignerOrProviderLike
+): Signer | Provider => {
+  if (typeof signerOrProvider === 'string') {
+    return new ethers.providers.JsonRpcProvider(signerOrProvider)
+  } else if (Provider.isProvider(signerOrProvider)) {
+    return signerOrProvider as Provider
+  } else if (Signer.isSigner(signerOrProvider)) {
+    return signerOrProvider as Signer
   } else {
     throw new Error('Invalid provider')
   }
