+++ packages/sdk/test/utils/coercion.spec.ts
@@ -3,17 +3,17 @@ import { Contract } from 'ethers'
 import { ethers } from 'hardhat'
 
 import { expect } from '../setup'
-import { toProvider, toTransactionHash } from '../../src'
+import { toSignerOrProvider, toTransactionHash } from '../../src'
 
 describe('type coercion utils', () => {
-  describe('toProvider', () => {
+  describe('toSignerOrProvider', () => {
     it('should convert a string to a JsonRpcProvider', () => {
-      const provider = toProvider('http://localhost:8545')
+      const provider = toSignerOrProvider('http://localhost:8545')
       expect(Provider.isProvider(provider)).to.be.true
     })
 
     it('should not do anything with a provider', () => {
-      const provider = toProvider(ethers.provider)
+      const provider = toSignerOrProvider(ethers.provider)
       expect(provider).to.deep.equal(ethers.provider)
     })
   })
